Added projects I did during that course on Udemy
In Health Potion I created a simple program that simulates healing potion in video game and, upon usage, adds random amount of health, depending on difficulty of the game.

Email Slicer project were created for a simple program that takes users email address and slices out username and domain name.

Travis is a project in which in the program were created a robot named Travis that knows a list of users and, when launched, it would ask user for a name and if that name isn't in the list, than user can add it to the list. If the name is on the list, than user can either remove it from there or leave it as it is.

Cinema simulator is a project in which the dictionary with movies were created and the user are being asked what movie user would like to watch. If the name of the movie is in the dictionary, user are being asked about their age (if user is too young, corresponding message appears) and program checks for number of available seats. If there is enough free seats, program remembers that there's one less seat for that film and prints message "Enjoy the film!". If there is not enough seats, then it says that they're sold out.

Baby project is about program that simulates baby behavior in asking some random question and, after answering it, it just always asks "Why?", until user types "just because", after which loop is over.

Pig latin is an interesting project, because it converts users message into pig latin.

Game project is just an ordinary Tic-Tac-Toe game in Python with board and two players.

Bank project is a program that utilizes classes in Python to create a simulation of a bank, where user can deposit money, withdraw it or see current balance.